Fix Error on No Published Questions and Minor Formatting
Was throwing error if viewer clicks "View Questions (Viewer)". Problems was bad SQL
text asking if a boolean column equaled true, when should have just been the column.
Unfortunately, now a page with no questions but the instruction to "click any question
to view it" when it should inform the reader that there are currently no questions
published.

A question that has been answered substantively (i.e., more than just a few words)
did not get wrapped in the div block and ran off the right of the page. Don't know
why the <pre> tag was used for the question-answer in the first place, but I deleted
that and added a line-break.